writeback: Do not sort b_io list only because of block device inode
It is very likely that block device inode will be part of BDI dirty list
as well. However it doesn't make sence to sort inodes on the b_io list
just because of this inode (as it contains buffers all over the device
anyway). So save some CPU cycles which is valuable since we hold relatively
contented wb->list_lock.

Change-Id: I69b2206bbbc870332c972ea0a64180c318608ef6
Signed-off-by: Jan Kara <jack@suse.cz>