ion: Actually check start exists
For whatever reason, start and offset are zero when using the OnePlus
camera or screenshotting, triggering the -EINVAL return. This causes
some crazy artifacts and distortion in images. As far as I can tell, a
zero vaddr is perfectly fine according to ion_no_page_cache_ops so I
believe this was removed in error.

Fixes: 09874396dfbf ("ion: ensure CMO target is valid")
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>